backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
yfinance
fredapi
feedparser
openai
httpx
numpy
pandas
ta
fpdf2
aiosqlite
